fix: ignore labelをpredictionにも適応

+++ src/train_val.py
@@ -73,15 +73,19 @@ def val(device, model, val_progress_bar, criterion, evaluator):
 
         if label.dim() == 4:
             label = label.squeeze(1)
+        
+        # ラベルをLong型に変換
+        label = label.long()
 
         with torch.no_grad():
             y = model(image)
 
-        loss = criterion(y, label.long())
+        loss = criterion(y, label)  # ignore_labelは自動的に無視される
         pred = get_pred(y)
         pred = pred.data.cpu().numpy()
         label = label.cpu().numpy()
-        evaluator.add_batch(label, pred)
+        
+        evaluator.add_batch(label, pred)  # Evaluator内でignore_labelが処理される
         val_progress_bar.set_postfix({'loss': f'{loss.item():.4f}'})
 
     mIoU = evaluator.Mean_Intersection_over_Union()
@@ -100,15 +104,19 @@ def test(cfg, device, model, test_loader, criterion, evaluator):
 
         if label.dim() == 4:
             label = label.squeeze(1)
+        
+        # ラベルをLong型に変換
+        label = label.long()
 
         with torch.no_grad():
             output = model(image)
         
-        loss = criterion(output, label.long())
+        loss = criterion(output, label)  # ignore_labelは自動的に無視される
         pred = get_pred(output)
         pred = pred.data.cpu().numpy()
         label = label.cpu().numpy()
-        evaluator.add_batch(label, pred)
+        
+        evaluator.add_batch(label, pred)  # Evaluator内でignore_labelが処理される
         test_progress_bar.set_postfix({'loss': f'{loss.item():.4f}'})
 
     mIoU = evaluator.Mean_Intersection_over_Union()
